Remove duplicated pagination class for Bootstrap
The `.pagination` class at Bootstrap adds margin at the top and at
the bottom of the element it is on. Looking at the [current examples][0]
they have on their documentation page, only one element has this class
and it’s an inner one, not the parent `<nav>`.

While working on one of our projects, I noticed that Pagy had way more
vertical margins than it should have, so I noticed its Bootstrap helper
was adding the class twice on two different elements. Since I believe the
markup it outputs should be as close as possible to Bootstrap’s
documentation, the class added to the `<nav>` element feels out of place.

[0]: https://getbootstrap.com/docs/4.4/components/pagination/#overview