add items_per_page set to original value of items
I ran into an astonishing issue. I was passing the Pagy instance to my custom pagination component and using its `items` attribute to set the selected option of a dropdown to select items per page. I happened to filter my data so that there were only 5 results, and the select changed to 5. I then discovered that `items` gets adjusted to the number of items when on the last page (when there were only 5 results, there was only one page, so it was on the last page).

As a workaround, I've had to explicitly pass the `items` value separately to my pagination component. However it would be nice if Pagy kept the original value of `items` somewhere so that it can be used for purposes similar to my own.

This PR is my suggestion as to how this could be achieved without any breaking changes.