Ruby 3 compatibility for Searchkick extra
When using Pagy together with Searchkick in Ruby 3, there is an
ArgumentError raised. This is because Pagy::Backend#pagy_searchkick is
using an options hash as last argument, while Searchkick::Model#search
expects keyword arguments. In previous Ruby versions, Ruby did some
magic to auto-convert the last argument to keyword args if it is an
Hash. This is no longer the case in Ruby 3, causing the ArgumentError
mentioned above.

The failure can be seen in this minimal Rails application:
https://github.com/tosch/pagy-searchkick-minimal-rails

This fix will ensure keyword arguments are being used when running in
Ruby 3, but will rely on the old behaviour for previous Ruby versions.

With those changes applied, the above minimal Rails application works as
expected.