fix: add functional tests for rails script + helpers
Why do we need this?
  Because, as a rails user, when debugging #341 (https://github.com/ddnexus/pagy/issues/341)
I wanted to see a functional test - because it is familiar to rails
users, and makes it super easy to debug.

Also small additional changes have been made to facilitate this - byebug
as a debugging tool, and I have copied the script used by rails and made
minor modifications where appropriate.

fix: debug script in line with discussions in #342