update: to test pagination with real data rows
Please refer to [issue](https://github.com/ddnexus/pagy/issues/356)
In order to reproduce it, some additional code was required - e.g.
setting up connections, adding sqllite, with examples to allow for
the easy reproduction of issues.

Hopefully, with this PR, we will see the perfect reproduction of issues.

I hope this helps.