add: github action which builds and deploys documentation
Adds a github action which:

(i) builds the pagy documentation site and
(ii) deploys it.

It builds the documentation on the `docs-site` branch.

This branch is then deployed via github pages.

Any PRs made on the master branch will trigger the docs to be
rebuilt. I am not sure whether non-documentation changes will
trigger the build.

Please see the following image