Fix the args in the Pagy::Countless#series override
When extras/countless is used in combination with extras/bootstrap under
Ruby 3 and the method `#pagy_bootstrap_combo_nav_js` is
called, an error was raised in lib/pagy.rb:47:in `series':

`ArgumentError: wrong number of arguments (given 1, expected 0)`

This is due to how Ruby 3 separates positional and keyword args. This
fix allows to forward properly both kinds of args to `super`.