Fix links to docs
These links are 404. They should either be .md or as in this commit with link to the pagy website.

Agreed: relative links used with .md extension. Why?

* We would want the docs to work from many domains
e.g. from your own fork of github pages, or from a
development environment. hard-coding to the main repo, while
it does work, is somewhat restrictive.